+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   "evebox.archived" and "evebox.escalated". This should not be noticed
   as EveBox has been using both tags for a very long.
 
+### Breaking Changes
+- Remove adding of Suricata rules to events on import. Suricata can
+  now be configured to include the rule in an alert.
+
 ## 0.11.0 - 2020-03-26
 
 ### Enhancements